turret subsystem
part of the bare-bones of the class is in. its definitily not gonna work in its current state, plus i have to set up wrap around so that it goes the direction closest to where it needs to be.

also, random error that i dont know what its about, it appeared as soon as i created the class but there was literally nothing except for a blank constructor.... idk:

Expected unqualified-idclang(expected_unqualified_id)

thats the error id.